+++ src/NumberTransformer/NumberTransformer.php
@@ -9,5 +9,10 @@ interface NumberTransformer
     /**
      * @throws NumberToWordsException
      */
-    public function toWords(int $number): string;
+    public function toWords(float $number): string;
+
+    /**
+     * @throws NumberToWordsException
+     */
+    // public function toWordsDecimal(float $number): string;
 }
